refactor: reorganize network/protocol
